(GP-2.1.1): Repository Initialization Testing
Updated Git.java and added a feature to the initializeRepo() method that verifies repo initialization by calling the new verifyRepoExists() method. Additionally, the resetRepo() method was added, which deletes repositories. Finally, tests were added to verify that resetRepo() works, the README was updated, and bug fixes in initializeRepo() were done.