(GP-2.1): Program Can Initialize Repository
* (GP-2.1): Program Can Initialize Repository

Created an initializeRepo() method that initializes the repository with a main directory, and 3 sub directories called objects, index, and HEAD. If the repo exists, nothing new is created. Updated the documentation in the README file.

* (GP-2.1.1): Repository Initialization Testing

Updated Git.java and added a feature to the initializeRepo() method that verifies repo initialization by calling the new verifyRepoExists() method. Additionally, the resetRepo() method was added, which deletes repositories. Finally, tests were added to verify that resetRepo() works, the README was updated, and bug fixes in initializeRepo() were done.