(GP-2.3.1) BLOB Files Confirmed
Verified that the BLOB file was created in the objects directory in the createBlob() method. This method now returns true if the file was created, and false if not. Additionally, the tester was updated to test if the verification works, and a resetObjectFiles() method was created in order to retest the createBlob() method. Finally, the README.md was updated once more.