(GP-2.3.2): BLOB Compression
BLOB files are now compressed before they are added to the objects directory. This was done by creating a new method called compressFile(). Additionally, compression can be toggled on and off. The tester was then updated to test the new compression feature. Finally, the README.md was updated, and minor bug fixes were done to the initializeRepository() method.