(GP-2.3.2): BLOB Compression (#3)
* (GP-2.3) Create BLOB Files

Created a method called createBlob() that creates a file in the objects directory, with its name being the SHA1 hash associated with the contents inside the file. Added a tester to test whether or not this works, and updated the README.md file.

* (GP-2.3.1) BLOB Files Confirmed

Verified that the BLOB file was created in the objects directory in the createBlob() method. This method now returns true if the file was created, and false if not. Additionally, the tester was updated to test if the verification works, and a resetObjectFiles() method was created in order to retest the createBlob() method. Finally, the README.md was updated once more.

* (GP-2.3.2): BLOB Compression

BLOB files are now compressed before they are added to the objects directory. This was done by creating a new method called compressFile(). Additionally, compression can be toggled on and off. The tester was then updated to test the new compression feature. Finally, the README.md was updated, and minor bug fixes were done to the initializeRepository() method.