(GP-2.4) Update Index File
Created the updateIndex() method, which updates the index file to display the SHA1 hash name of the BLOB file, as well as the name of the file next to it. I then created a tester which shows that this method works as intended. Finally, I updated the README.md file.