(GP-2.4.2) Reset All Files (#4)
* (GP-2.4) Update Index File

Created the updateIndex() method, which updates the index file to display the SHA1 hash name of the BLOB file, as well as the name of the file next to it. I then created a tester which shows that this method works as intended. Finally, I updated the README.md file.

* (GP-2.4.1) Comprehensive Tester

So....... I might have already accidentally done this in (GP-2.4).

* (GP-2.4.2) Reset All Files

Created a method in the tester called resetEverything() which resets the entire repository to it's basic files (main directory, empty objects directory, empty index and HEAD files). Made some bug fixes and updated the tester.