(GP-3.3): Finished developing createIndexTree() (#7)
* (GP-3.3.1) Started creating the IndexTree

Moved tree methods to a tree class, created createIndexTree()

* (GP-3.3.2): Fixed bugs in createTree()

fixed some bugs that were made in commit (GP-3.2)

* (GP-3.3.3) Fixed everything for createTree()

createTree() now functions as expected

* (GP-3.3.4) Fixed bugs with updateIndex()

updateIndex() was not working as intended, and so I fixed the method. It now correctly updates the index and functions much smoother.

* (GP-3.3.5): Finished developing createIndexTree()

The method createIndexTree() successfully creates a tree based on the index file in the git repo. This method calls upon indexTreeHelper() which constructs the tree, and greatestLeaf() which returns the leaf-most directory. Everything functions as expected.