(GP-5.0): Rubric (#8)
* (GP-5.1): Added tests for init() and add()

Created tests for the GitWrapper methods of init() and add(). I also went into the Git class and fixed the updateIndex() method, which was previously written incorrectly.

* (GP-5.2): Updated the  tester format

Cleaned up the tester file and re-added all my previous tests. Deleted the compression method.

* (GP-5.3): Finished testing trees, updated createTree()

createTree() was not correctly generating the tree, so I fixed this and now it works as expected. Finished creating the tester.